requirements.txt
streamlit
PyMuPDF
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
